move insist.js up a level, out of src/kernel/ into src/
