+++ src/insist.js
