+++ src/kernel/deviceSlots.js
@@ -1,7 +1,7 @@
 import harden from '@agoric/harden';
 import Nat from '@agoric/nat';
 import { QCLASS, mustPassByPresence, makeMarshal } from '@agoric/marshal';
-import { insist } from './insist';
+import { insist } from '../insist';
 import {
   insistVatType,
   makeVatSlot,
