+++ src/kernel/parseKernelSlots.js
@@ -1,5 +1,5 @@
 import Nat from '@agoric/nat';
-import { insist } from './insist';
+import { insist } from '../insist';
 
 // Object/promise references (in the kernel) contain a two-tuple of (type,
 // index). All object references point to entries in the kernel Object
