+++ src/kernel/state/deviceKeeper.js
@@ -1,5 +1,5 @@
 import harden from '@agoric/harden';
-import { insist } from '../insist';
+import { insist } from '../../insist';
 import { parseKernelSlot } from '../parseKernelSlots';
 import { makeVatSlot, parseVatSlot } from '../../vats/parseVatSlots';
