+++ src/vats/comms/parseRemoteSlot.js
@@ -1,5 +1,5 @@
 import Nat from '@agoric/nat';
-import { insist } from '../../kernel/insist';
+import { insist } from '../../insist';
 
 // Object/promise references (in remote messages) contain a three-tuple of
 // (type, allocator flag, index). The allocator flag inside an inbound
